Avoid analysis dependence on opt by splitting some of OptimizerUtils into AnalysisUtils
